Split child document E2E coverage
